server/requirements.txt
websockets
python-environ
redis
hiredis
pyjwt